+++ packages/expo/src/Expo.fx.web.tsx
@@ -1,5 +1,3 @@
-import { setupDomProcessEnv } from './dom/setup-dom-base-url';
-
 // When users dangerously import a file inside of react-native, it breaks the web alias.
 // This is one of the most common, and cryptic web errors that users encounter.
 // This conditional side-effect provides a more helpful error message for debugging.
@@ -18,5 +16,3 @@ if (__DEV__) {
     });
   }
 }
-
-setupDomProcessEnv();