+++ packages/babel-preset-expo/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Uses `process.env.EXPO_DOM_BASE_URL` to reference the DOM Components root URL. ([#32502](https://github.com/expo/expo/pull/32502) by [@kudo](https://github.com/kudo))
+
 ## 12.0.0-preview.4 — 2024-10-31
 
 ### 🐛 Bug fixes